+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## main
+
+### ✨ Features and improvements
+
+- Write clustered pmtiles, add debug pipeline (#18)
+
 ## 0.0.1
 
 ### ✨ Features and improvements